+++ src/main/java/frc/utils/joysticks/JoystickPorts.java
@@ -1,0 +1,21 @@
+package frc.utils.joysticks;
+
+public enum JoystickPorts {
+
+    MAIN(0),
+    SECOND(1),
+    THIRD(2),
+    FOURTH(3),
+    FIFTH(4),
+    SIXTH(5);
+
+    private final int port;
+
+    JoystickPorts(int port) {
+        this.port = port;
+    }
+
+    public int getPort() {
+        return port;
+    }
+}
